add haralick features
