+++ src/main/java/net/imagej/ops/features/sets/Haralick3DFeatureSet.java
@@ -30,8 +30,12 @@
 
 package net.imagej.ops.features.sets;
 
+import java.util.HashSet;
+import java.util.Set;
+
 import net.imagej.ops.Contingent;
-import net.imagej.ops.features.AutoResolvingFeatureSet;
+import net.imagej.ops.OpRef;
+import net.imagej.ops.features.AbstractAutoResolvingFeatureSet;
 import net.imagej.ops.features.FeatureSet;
 import net.imagej.ops.features.firstorder.FirstOrderFeatures.MaxFeature;
 import net.imagej.ops.features.firstorder.FirstOrderFeatures.MinFeature;
@@ -63,12 +67,14 @@ import org.scijava.plugin.Plugin;
  * {@link FeatureSet} containing Haralick texture {@link Feature}s
  * 
  * @author Christian Dietz (University of Konstanz)
+ * @author Andreas Graumann (University of Konstanz)
  * 
  * @param <I>
  */
 @Plugin(type = FeatureSet.class, label = "Haralick Features")
 public class Haralick3DFeatureSet<T> extends
-AutoResolvingFeatureSet<IterableInterval<T>, DoubleType> implements Contingent {
+		AbstractAutoResolvingFeatureSet<IterableInterval<T>, DoubleType> implements
+		Contingent {
 
 	@Parameter(type = ItemIO.INPUT, label = "Number of Gray Levels", description = "Determines the size of the co-occurrence matrix", min = "1", max = "2147483647", stepSize = "1")
 	private double nrGrayLevels = 8;
@@ -82,7 +88,7 @@ AutoResolvingFeatureSet<IterableInterval<T>, DoubleType> implements Contingent {
 			"DIAGONAL_VERTICAL", "DIAGONAL_DIAGONAL", "ANTIDIAGONAL_VERTICAL",
 			"ANTIDIAGONAL_DIAGONAL", "DEPTH" })
 	private String orientation;
-	
+
 	public void setDistance(final double distance) {
 		this.distance = distance;
 	}
@@ -107,31 +113,6 @@ AutoResolvingFeatureSet<IterableInterval<T>, DoubleType> implements Contingent {
 		this.nrGrayLevels = nrGrayLevels;
 	}
 
-	public Haralick3DFeatureSet() {
-
-		addOutputOp(ASMFeature.class);
-		addOutputOp(ClusterPromenenceFeature.class);
-		addOutputOp(ClusterShadeFeature.class);
-		addOutputOp(ContrastFeature.class);
-		addOutputOp(DifferenceVarianceFeature.class);
-		addOutputOp(DifferenceEntropyFeature.class);
-		addOutputOp(EntropyFeature.class);
-		addOutputOp(ICM1Feature.class);
-		addOutputOp(ICM2Feature.class);
-		addOutputOp(IFDMFeature.class);
-		addOutputOp(SumAverageFeature.class);
-		addOutputOp(SumEntropyFeature.class);
-		addOutputOp(SumVarianceFeature.class);
-		addOutputOp(VarianceFeature.class);
-		addOutputOp(TextureHomogenityFeature.class);
-		addOutputOp(MaxProbabilityFeature.class);
-
-		// add cooc parameters
-		addHiddenOp(CooccurrenceMatrix3D.class, getInput().getClass(),
-				nrGrayLevels, distance, orientation, MinFeature.class,
-				MaxFeature.class);
-	}
-
 	@Override
 	public boolean conforms() {
 
@@ -141,5 +122,37 @@ AutoResolvingFeatureSet<IterableInterval<T>, DoubleType> implements Contingent {
 		}
 
 		return count == 3;
+	}
+
+	@Override
+	public Set<OpRef<?>> getOutputOps() {
+		
+		final HashSet<OpRef<?>> outputOps = new HashSet<OpRef<?>>();
+		outputOps.add(createOpRef(ASMFeature.class));
+		outputOps.add(createOpRef(ClusterPromenenceFeature.class));
+		outputOps.add(createOpRef(ClusterShadeFeature.class));
+		outputOps.add(createOpRef(ContrastFeature.class));
+		outputOps.add(createOpRef(DifferenceVarianceFeature.class));
+		outputOps.add(createOpRef(DifferenceEntropyFeature.class));
+		outputOps.add(createOpRef(EntropyFeature.class));
+		outputOps.add(createOpRef(ICM1Feature.class));
+		outputOps.add(createOpRef(ICM2Feature.class));
+		outputOps.add(createOpRef(IFDMFeature.class));
+		outputOps.add(createOpRef(SumAverageFeature.class));
+		outputOps.add(createOpRef(SumEntropyFeature.class));
+		outputOps.add(createOpRef(SumVarianceFeature.class));
+		outputOps.add(createOpRef(VarianceFeature.class));
+		outputOps.add(createOpRef(TextureHomogenityFeature.class));
+		outputOps.add(createOpRef(MaxProbabilityFeature.class));
+		
+		return outputOps;
+	}
+
+	@Override
+	public Set<OpRef<?>> getHiddenOps() {
+		final HashSet<OpRef<?>> hiddenOps = new HashSet<OpRef<?>>();
+		hiddenOps.add(createOpRef(CooccurrenceMatrix3D.class, getInput(), nrGrayLevels,
+				distance, orientation, MinFeature.class, MaxFeature.class));
+		return hiddenOps;
 	}
 }
